Work on front-end. Complete restyling.

+++ templates/list.php
@@ -17,17 +17,20 @@ ob_start();
                 <tbody>
                     
                         <!-- Boucle affichage : récupération des lignes de la table post-->
-                <?php foreach ( $billets_admin as $post):
+                <?php if (!empty($billets_admin)) {
+                    foreach ( $billets_admin as $post):
                     {
                         echo '<tr> <td>' . $post->getTitre() . '</td><td>' . $post->getDatecreation() .
-                        '</td><td>' . $post->getDatemaj() . '</td>
+                            '</td><td>' . $post->getDatemaj() . '</td>
                         <td><div class="btn-group">
                             <a class="btn btn-primary btn-sm" href="/admin/updatepost/' . $post->getId() . ' " >Modifier</a>
                             <a class="btn btn-danger btn-sm"  href="/admin/delete/' . $post->getId() . ' " >Supprimer</a>
                             </div> 
                         </td> </tr> <br/>';
                     }
-                  endforeach;
+                    endforeach;
+                } else
+                    echo '<h3>Pas de billet dans la base de données</h3>';
                 ?>
                         <!-- Fin boucle -->
